Download logs
WIP for downloading the logs instead of copying them. This is needed as the logs output might be too large to be handled by the clipboard.

The current solution works in the browser, but does not work on Android as the file doesn't get saved to the download folder.

According to [this bug report](https://github.com/dotnet/maui/issues/15493), though the code is mostly adapted from the [official sample](https://learn.microsoft.com/en-us/aspnet/core/blazor/file-downloads?view=aspnetcore-8.0).